+++ Annex/src/utopia/annex/controller/PersistedRequestHandler.scala
@@ -1,6 +1,6 @@
 package utopia.annex.controller
 
-import utopia.annex.model.request.ApiRequest
+import utopia.annex.model.request.{ApiRequest, ApiRequestSeed, RequestQueueable}
 import utopia.annex.model.response.RequestResult
 import utopia.flow.generic.factory.FromModelFactory
 import utopia.flow.generic.model.immutable.Model
@@ -17,8 +17,9 @@ trait PersistedRequestHandler
 	/**
 	  * @return Factory used for parsing requests.
 	  *         Only needs to process models accepted by 'shouldHandle'.
+	  *         Returns the requests in either prepared form or in "seed" form
 	  */
-	def factory: FromModelFactory[ApiRequest]
+	def factory: FromModelFactory[Either[ApiRequestSeed, ApiRequest]]
 	
 	/**
 	  * @param requestModel A model
@@ -28,8 +29,9 @@ trait PersistedRequestHandler
 	/**
 	  * Handles a response received for a persisted request
 	  * @param requestModel The persisted model from which the request was parsed
-	  * @param request The request that was parsed and (possibly) sent to the server
+	  * @param request The request that was parsed and (possibly) sent to the server.
+	  *                Either in prepared form (Right) or in "seed" form (Left).
 	  * @param result Result received for the persisted request
 	  */
-	def handle(requestModel: Model, request: ApiRequest, result: RequestResult): Unit
+	def handle(requestModel: Model, request: RequestQueueable, result: RequestResult): Unit
 }